test/requirements.txt
simplejson
pyaml
requests
qubell-api-python-client>=1.23.24.1
elasticsearch
